refactor(exporters): promote 3-layer R2 exporters to default (PR-5/7 commit 6)
PR-4 で並行運用していた from-indicators / from-observations exporter を
正本 (ranking-items-snapshot.ts / ranking-values-snapshot.ts) に昇格。

- ranking-items-snapshot.ts: indicators + indicator_tags JOIN による snapshot 生成
- ranking-values-snapshot.ts: observations + indicators JOIN による partition 出力
- exporters/index.ts: from-* バリアント export を削除
- 削除: ranking-items-from-indicators-snapshot.ts / ranking-values-from-observations-snapshot.ts (重複コード)
- 削除: scripts/diff-snapshots-old-vs-new.ts / scripts/export-snapshots-from-3layer.ts (PR-5 完了で不要)

R2 出力 JSON は完全互換（PR-4 で byte-level diff 検証済み）。
